chore: remove build and IDE cache folders from repo
