+++ RERPAPI/.gitignore
@@ -3,3 +3,14 @@
 ################################################################################
 
 bin
+
+# Bỏ qua các file/thư mục build
+bin/
+obj/
+
+# Bỏ qua file user-specific
+*.user
+*.suo
+
+# Visual Studio specific
+.vs/
